+++ src/coin-selection/lib/Convex/CoinSelection/Class.hs
@@ -12,8 +12,7 @@ module Convex.CoinSelection.Class(
   runTracingBalancingT
 ) where
 
-import           Cardano.Api.Shelley              (AddressInEra, BabbageEra,
-                                                   BuildTx, TxBodyContent)
+import           Cardano.Api.Shelley              (AddressInEra, BabbageEra)
 import qualified Cardano.Api.Shelley              as C
 import           Control.Monad.Catch              (MonadCatch, MonadMask,
                                                    MonadThrow)
@@ -25,14 +24,15 @@ import           Control.Monad.Trans.Class        (MonadTrans (..))
 import qualified Control.Monad.Trans.State        as StrictState
 import qualified Control.Monad.Trans.State.Strict as LazyState
 import           Control.Tracer                   (Tracer, natTracer)
+import           Convex.BuildTx                   (TxBuilder)
 import           Convex.Class                     (MonadBlockchain (..),
                                                    MonadMockchain (..))
-import           Convex.Query                     (MonadUtxoQuery(utxosByPaymentCredentials))
 import           Convex.CoinSelection             (BalanceTxError,
                                                    TxBalancingMessage)
 import qualified Convex.CoinSelection
 import           Convex.CardanoApi.Lenses         (emptyTxOut)
 import           Convex.MonadLog                  (MonadLog, MonadLogIgnoreT)
+import           Convex.Query                     (MonadUtxoQuery (utxosByPaymentCredentials))
 import           Convex.Utxos                     (BalanceChanges (..),
                                                    UtxoSet (..))
 
@@ -60,7 +60,7 @@ class Monad m => MonadBalance m where
     UtxoSet C.CtxUTxO a ->
 
     -- | The unbalanced transaction body
-    TxBodyContent BuildTx BabbageEra ->
+    TxBuilder ->
 
     -- | The balanced transaction body and the balance changes (per address)
     m (Either BalanceTxError (C.BalancedTxBody BabbageEra, BalanceChanges))